requirements.txt
numpy
matplotlib
torch
jupyter
tiktoken
python-dotenv
